+++ pkg/grafana/client/factory.go
@@ -100,8 +100,9 @@ func (g *DefaultGrafanaClientGenerator) GenerateGrafanaClient(ctx context.Contex
 		// This overrides the server side defined org context,
 		// see https://grafana.com/docs/grafana/latest/developers/http_api/user/#switch-user-context-for-signed-in-user
 		// This ensures operations like deleting organizations other than the first org. work as expected.
-		OrgID:      1,
-		Host:       grafanaURL.Host,
+		OrgID: 1,
+		Host:  grafanaURL.Host,
+		// TODO using a serviceaccount later would be better as they are scoped to an organization
 		BasicAuth:  url.UserPassword(string(adminUsernameBytes), string(adminPasswordBytes)),
 		NumRetries: clientConfigNumRetries,
 		TLSConfig:  clientTLSConfig,