+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Added a `generateDatasources` to generate all datasource needed for an organization
+- Major refactoring of the `ConfigureDefaultDatasources` method into `ConfigureDatasource`
+- Remove explicit logic to delete `gs-mimir-old` datasource (now covered)
+- Optimize GrafanaOrganization status update to only be performed when necessary
+
 ## [0.41.0] - 2025-09-01
 
 ### Added